#75 refactoring: fix conflict
