+++ src/main/java/com/example/algoproject/errors/exception/badrequest/AlreadyExistSolutionException.java
@@ -1,4 +1,4 @@
-package com.example.algoproject.errors.exception;
+package com.example.algoproject.errors.exception.badrequest;
 
 public class AlreadyExistSolutionException extends RuntimeException{
     private static final String MESSAGE = "이미 해당 문제에 솔루션이 존재합니다.";
